haven't commitied in ages commit

+++ StockMarket.AccountAPI/Models/IPO.cs
@@ -26,7 +26,7 @@ namespace StockMarket.AccountAPI.Models
         [Column(TypeName = "Date")]
         public DateTime OpenDate { get; set; }
         [Required]
-        public int PricePerShare { get; set; }
+        public double PricePerShare { get; set; }
         [Required]
         public int TotalNumberOfShares { get; set; }
         [StringLength(150)]
